Atualizando bibliotecas jars (trove 3.0.3) Aplicando Safe Admin System, Bot Report Button, Bloqueio de IP. Corrigindo protecoes antigas. Skills Ekimus, Mobs Hall of Suffering(Rotten Messenger, Zealot of Infinity, Yehan klodekus, Yehan Klanikus, Soul Exploiter, etc)  revistas e corrigidas Corrigindo Stats de Energy Seeds Npcs, Destroyed Tumor, Tumor of Death, e mobs de HallofSuffering em geral

+++ CT23-DataPack/data/scripts/handlers/admincommandhandlers/AdminDelete.java
@@ -36,6 +36,9 @@ public class AdminDelete implements IAdminCommandHandler
 	
 	public boolean useAdminCommand(String command, L2PcInstance activeChar)
 	{
+		if (activeChar == null || !activeChar.getPcAdmin().canUseAdminCommand())
+			return false;
+		
 		if (command.equals("admin_delete"))
 			handleDelete(activeChar);
 		return true;
